Created New HealthBar UI and anchor progress bar to the entire canvas. Then create binding of progress bar. Create a new variable MyTank. In Tank class, create GetHealthPercent () const which returns CurrentHealth / StartingHealth casted as float so percentage division will be more accurate in decimal place. Then , set the method as UFUNCTION(BlueprintPure) because BlueprintPure is the equilavent of const in Blueprints.
In Tank_BP, Create a new Widget in hierachy, then try and Set MyTank in HeathBar UI by Get Widget , wire to Get User Widget wire to Cast to HealthBar which As HealthBar wires to Set MyTank , use a self node and connect to MyTank as a reference. In viewport, adjust Widget above and tank and set its pixels smaller , then set it as Screen.

Now, Tank will have healthbars above them that will reduce when they are hit, and will be depleted after five hits.